// web/src/components/Appearance/interfaces.ts
export interface CameraState {
  head: boolean;
  body: boolean;
  bottom: boolean;
}

export interface ClothesState {
  head: boolean;
  body: boolean;
  bottom: boolean;
}

export interface RotateState {
  left: boolean;
  right: boolean;
}